test: Enable `ContextFlyout` and `SelectionFlyout` tests on WinUI
